# src/main/resources/static/openapi/eventx-api.yaml
openapi: 3.0.3
info:
  title: EventX Manager API
  version: 2.1.0
  description: |
    API REST per la gestione della compravendita di biglietti per eventi su territorio nazionale.
    Espone endpoint per **Utenti**, **Eventi**, **Location**, **Recensioni** e **Biglietti**.

    ## Autenticazione
    Tutti gli endpoint (eccetto registrazione e login) richiedono un token JWT nell'header:
    ```
    Authorization: Bearer <token>
    ```
    Il token si ottiene chiamando `POST /api/v1/utenti/login` o `POST /api/v1/utenti/login/username`.
servers:
  - url: /
    description: Server corrente (locale o Railway)

# Sicurezza globale: tutti gli endpoint la ereditano salvo override esplicito
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/utenti:
    post:
      tags: [Users]
      summary: Registra un nuovo utente
      security: []   # endpoint pubblico — nessun token richiesto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
            example:
              username: mario_rossi
              email: mario.rossi@example.com
              password: MySecret123
      responses:
        '201':
          description: Utente creato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags: [Users]
      summary: Lista utenti (filtro opzionale per username)
      parameters:
        - in: query
          name: cerca
          description: Ricerca parziale case-insensitive sullo username
          schema:
            type: string
          example: mario
      responses:
        '200':
          description: Lista utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
  /api/v1/utenti/login:
    post:
      tags: [Users]
      summary: Login tramite email e password
      security: []   # endpoint pubblico — nessun token richiesto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
            example:
              email: mario.rossi@example.com
              password: MySecret123
      responses:
        '200':
          description: Login effettuato — restituisce token JWT e profilo utente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/utenti/login/username:
    post:
      tags: [Users]
      summary: Login tramite username e password
      security: []   # endpoint pubblico — nessun token richiesto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameLoginRequestDTO'
            example:
              username: mario_rossi
              password: MySecret123
      responses:
        '200':
          description: Login effettuato — restituisce token JWT e profilo utente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/utenti/{id}:
    get:
      tags: [Users]
      summary: Dettaglio utente per ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Utente trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Elimina utente
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Utente eliminato
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/utenti/profilo:
    post:
      tags: [ Users ]
      summary: Aggiorna i dati del profilo utente
      description: |
        Modifica email, nome, cognome, telefono e email PayPal.
        Richiede l'**ID utente**, la **password corrente** e la nuova **email**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequestDTO'
            example:
              userId: "user_abc123"
              password: "MySecret123"
              email: "mario.nuovo@example.com"
              firstName: "Mario"
              lastName: "Rossi"
              phone: "3331234567"
              paypalEmail: "mario.nuovo@paypal.com"
      responses:
        '200':
          description: Profilo aggiornato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/utenti/{id}/password:
    patch:
      tags: [Users]
      summary: Modifica la password dell'utente
      description: |
        Richiede la **password corrente** e la **nuova password** (minimo 8 caratteri).
        Non restituisce corpo in caso di successo (204).
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDTO'
            example:
              currentPassword: "MySecret123"
              newPassword: "NewSecret456"
      responses:
        '204':
          description: Password modificata con successo
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/utenti/email/{email}:
    get:
      tags: [Users]
      summary: Dettaglio utente per email
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
          example: mario.rossi@example.com
      responses:
        '200':
          description: Utente trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── EVENTS ────────────────────────────────────────────────────────────────
  /api/v1/eventi:
    post:
      tags: [Events]
      summary: Crea un nuovo evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
            example:
              plannerId: "user_abc123"
              name: "Rock in Roma 2026"
              description: "Festival rock estivo"
              locationId: "loc_xyz789"
              date: "2026-07-15T21:00:00"
              duration: 180
              price: 45.00
              capacity: 5000
      responses:
        '201':
          description: Evento creato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Events]
      summary: Lista eventi con filtri opzionali
      security: []   # endpoint pubblico — nessun token richiesto
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/EventStatus'
        - in: query
          name: name
          description: Ricerca parziale case-insensitive sul nome
          schema:
            type: string
          example: Rock
        - in: query
          name: plannerId
          description: Filtra per ID utente creatore
          schema:
            type: string
        - in: query
          name: locationId
          description: Filtra per ID location
          schema:
            type: string
        - in: query
          name: dal
          schema:
            type: string
            format: date-time
          example: '2026-06-01T00:00:00'
        - in: query
          name: al
          schema:
            type: string
            format: date-time
          example: '2026-08-31T23:59:59'
      responses:
        '200':
          description: Lista eventi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDTO'
  /api/v1/eventi/{id}:
    get:
      tags: [Events]
      summary: Dettaglio evento per ID
      security: []   # endpoint pubblico — nessun token richiesto
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Evento trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Events]
      summary: Aggiorna evento
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
      responses:
        '200':
          description: Evento aggiornato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Events]
      summary: Elimina evento
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Evento eliminato
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/eventi/{id}/stato:
    patch:
      tags: [Events]
      summary: Aggiorna solo lo stato dell'evento
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - in: query
          name: status
          required: true
          schema:
            $ref: '#/components/schemas/EventStatus'
      responses:
        '200':
          description: Stato aggiornato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── LOCATIONS ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: ID MongoDB del documento
  responses:
    BadRequest:
      description: Richiesta non valida (validazione fallita)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Risorsa duplicata (email o username già in uso)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Token JWT mancante, scaduto o non valido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            errore: Non autorizzato
            messaggio: Token JWT non valido o scaduto
            timestamp: '2026-04-26T10:00:00'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT ottenuto tramite il login. Da inviare nell'header:
        `Authorization: Bearer <token>`
  schemas:
...
    EventStatus:
      type: string
      enum: [PUBLISHED, ACTIVE, ENDED]
      description: Stato del ciclo di vita dell'evento

    # ── USER ─────────────────────────────────────────────────────────────────
    LoginResponseDTO:
      type: object
      description: Risposta restituita dagli endpoint di login. Contiene il token JWT e i dati dell'utente.
      properties:
        token:
          type: string
          description: Token JWT da utilizzare nell'header Authorization per le successive chiamate
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          description: Tipo di token — sempre "Bearer"
          default: Bearer
          example: Bearer
        user:
          $ref: '#/components/schemas/UserResponseDTO'
    UserRequestDTO:
      type: object
      required: [username, email]
      properties:
        username:
          type: string
          example: mario_rossi
        email:
          type: string
          format: email
          example: mario.rossi@example.com
        password:
          type: string
          minLength: 8
          description: Obbligatoria in creazione; opzionale in aggiornamento (se assente la password non cambia)
          example: MySecret123
        paypalEmail:
          type: string
          format: email
          description: Email dell'account PayPal dell'utente, usata per pagamenti. Può differire dall'email di accesso.
          example: mario.rossi@paypal.com
    UserResponseDTO:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        paypalEmail:
          type: string
          format: email
          description: Email PayPal dell'utente
          example: mario.rossi@paypal.com
        isArtist:
          type: boolean
        isLocation:
          type: boolean
        activeTickets:
          type: array
          items:
            type: string
          description: ID dei biglietti attivi dell'utente
        lastTickets:
          type: array
          items:
            type: string
          description: ID dei biglietti scaduti/ceduti
        reviews:
          type: array
          items:
            type: string
          description: ID delle recensioni scritte dall'utente
        registrationDate:
          type: string
          format: date-time
    UpdateProfileRequestDTO:
      type: object
      required: [ userId, password, email ]
      description: "DTO per l'aggiornamento del profilo utente. I campi firstName, lastName, phone e paypalEmail sono opzionali."
      properties:
        userId:
          type: string
          description: ID dell'utente da aggiornare
          example: "user_abc123"
        password:
          type: string
          description: Password corrente — richiesta per confermare l'identità
          example: MySecret123
        email:
          type: string
          format: email
          description: Nuovo indirizzo email (obbligatorio)
          example: mario.nuovo@example.com
        firstName:
          type: string
          description: Nome dell'utente (opzionale)
          example: Mario
        lastName:
          type: string
          description: Cognome dell'utente (opzionale)
          example: Rossi
        phone:
          type: string
          description: Numero di telefono (opzionale)
          example: "3331234567"
        paypalEmail:
          type: string
          format: email
          description: Email account PayPal (opzionale)
          example: mario.nuovo@paypal.com
    ChangePasswordRequestDTO:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          description: Password attualmente in uso
          example: MySecret123
        newPassword:
          type: string
          minLength: 8
          description: Nuova password (minimo 8 caratteri)
          example: NewSecret456
    LoginRequestDTO:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: mario.rossi@example.com
        password:
          type: string
          example: MySecret123
    UsernameLoginRequestDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: mario_rossi
        password:
          type: string
          example: MySecret123

    # ── EVENT ─────────────────────────────────────────────────────────────────
    EventRequestDTO:
      type: object
      required: [plannerId, name, locationId, date, duration, capacity]
      properties:
        plannerId:
          type: string
          description: ID dell'utente che crea l'evento
        name:
          type: string
          example: Rock in Roma 2026
        description:
          type: string
        locationId:
          type: string
          description: ID della Location (documento separato nella collection locations)
        date:
          type: string
          format: date-time
          example: '2026-07-15T21:00:00'
        duration:
          type: integer
          minimum: 1
          description: Durata in minuti
          example: 180
        price:
          type: number
          format: double
          minimum: 0
          example: 45.00
        capacity:
          type: integer
          minimum: 1
          description: Capienza massima
          example: 5000
    EventResponseDTO:
      type: object
      properties:
        id:
          type: string
        plannerId:
          type: string
        name:
          type: string
        description:
          type: string
        locationId:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        price:
          type: number
          format: double
        capacity:
          type: integer
        expectedPublic:
          type: integer
        status:
          $ref: '#/components/schemas/EventStatus'
        reviews:
          type: array
          items:
            type: string
          description: ID delle recensioni associate all'evento
        registrationDate:
          type: string
          format: date-time

    # ── LOCATION ──────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        errore:
          type: string
          example: Not Found
        messaggio:
          type: string
          example: Risorsa non trovata con id - abc123
        timestamp:
          type: string
          format: date-time